WildFire01/requirements.txt
kagglehub>=0.2.7
tensorflow>=2.15
keras>=3.0
numpy>=1.23
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
joblib>=1.3
pyyaml>=6.0
tqdm>=4.66
